+++ src/PSRule/Runtime/Assert.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using Manatee.Json;
@@ -214,10 +214,11 @@ namespace PSRule.Runtime
                 return result;
 
             // Assert
-            for (var i = 0; i < prefix.Length; i++)
+            for (var i = 0; prefix != null && i < prefix.Length; i++)
+            {
                 if (value.StartsWith(prefix[i], caseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase))
                     return Pass();
-            
+            }
             return Fail(ReasonStrings.StartsWith, field, FormatArray(prefix));
         }
 
@@ -234,10 +235,11 @@ namespace PSRule.Runtime
                 return result;
 
             // Assert
-            for (var i = 0; i < suffix.Length; i++)
+            for (var i = 0; suffix != null && i < suffix.Length; i++)
+            {
                 if (value.EndsWith(suffix[i], caseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase))
                     return Pass();
-
+            }
             return Fail(ReasonStrings.EndsWith, field, FormatArray(suffix));
         }
 
@@ -254,11 +256,58 @@ namespace PSRule.Runtime
                 return result;
 
             // Assert
-            for (var i = 0; i < text.Length; i++)
+            for (var i = 0; text != null && i < text.Length; i++)
+            {
                 if (value.IndexOf(text[i], caseSensitive ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase) >= 0)
                     return Pass();
-
+            }
             return Fail(ReasonStrings.Contains, field, FormatArray(text));
+        }
+
+        /// <summary>
+        /// The object field value should only contain lowercase characters.
+        /// </summary>
+        public AssertResult IsLower(PSObject inputObject, string field, bool requireLetters = false)
+        {
+            // Guard parameters
+            if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
+                GuardNullOrEmptyParam(field, nameof(field), out result) ||
+                GuardField(inputObject, field, false, out object fieldValue, out result) ||
+                GuardString(fieldValue, out string value, out result))
+                return result;
+
+            for (var i = 0; i < value.Length; i++)
+            {
+                if (!char.IsLetter(value, i) && requireLetters)
+                    return Fail(ReasonStrings.IsLetter, value);
+
+                if (char.IsLetter(value, i) && !char.IsLower(value, i))
+                    return Fail(ReasonStrings.IsLower, value);
+            }
+            return Pass();
+        }
+
+        /// <summary>
+        /// The object field value should only contain uppercase characters.
+        /// </summary>
+        public AssertResult IsUpper(PSObject inputObject, string field, bool requireLetters = false)
+        {
+            // Guard parameters
+            if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
+                GuardNullOrEmptyParam(field, nameof(field), out result) ||
+                GuardField(inputObject, field, false, out object fieldValue, out result) ||
+                GuardString(fieldValue, out string value, out result))
+                return result;
+
+            for (var i = 0; i < value.Length; i++)
+            {
+                if (!char.IsLetter(value, i) && requireLetters)
+                    return Fail(ReasonStrings.IsLetter, value);
+
+                if (char.IsLetter(value, i) && !char.IsUpper(value, i))
+                    return Fail(ReasonStrings.IsUpper, value);
+            }
+            return Pass();
         }
 
         /// <summary>
@@ -283,7 +332,7 @@ namespace PSRule.Runtime
             return Pass();
         }
 
-        public AssertResult Greater(PSObject inputObject, string field, int value)
+        public AssertResult Greater(PSObject inputObject, string field, int value, bool convert = false)
         {
             // Guard parameters
             if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
@@ -291,13 +340,13 @@ namespace PSRule.Runtime
                 GuardField(inputObject, field, false, out object fieldValue, out result))
                 return result;
 
-            if (CompareNumeric(fieldValue, value, out int compare, out object actual))
+            if (CompareNumeric(fieldValue, value, convert, out int compare, out object actual))
                 return compare > 0 ? Pass() : Fail(ReasonStrings.Greater, actual, value);
 
             return Fail(ReasonStrings.Compare, fieldValue, value);
         }
 
-        public AssertResult GreaterOrEqual(PSObject inputObject, string field, int value)
+        public AssertResult GreaterOrEqual(PSObject inputObject, string field, int value, bool convert = false)
         {
             // Guard parameters
             if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
@@ -305,13 +354,13 @@ namespace PSRule.Runtime
                 GuardField(inputObject, field, false, out object fieldValue, out result))
                 return result;
 
-            if (CompareNumeric(fieldValue, value, out int compare, out object actual))
+            if (CompareNumeric(fieldValue, value, convert, out int compare, out object actual))
                 return compare >= 0 ? Pass() : Fail(ReasonStrings.GreaterOrEqual, actual, value);
 
             return Fail(ReasonStrings.Compare, fieldValue, value);
         }
 
-        public AssertResult Less(PSObject inputObject, string field, int value)
+        public AssertResult Less(PSObject inputObject, string field, int value, bool convert = false)
         {
             // Guard parameters
             if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
@@ -319,13 +368,13 @@ namespace PSRule.Runtime
                 GuardField(inputObject, field, false, out object fieldValue, out result))
                 return result;
 
-            if (CompareNumeric(fieldValue, value, out int compare, out object actual))
+            if (CompareNumeric(fieldValue, value, convert, out int compare, out object actual))
                 return compare < 0 ? Pass() : Fail(ReasonStrings.Less, actual, value);
 
             return Fail(ReasonStrings.Compare, fieldValue, value);
         }
 
-        public AssertResult LessOrEqual(PSObject inputObject, string field, int value)
+        public AssertResult LessOrEqual(PSObject inputObject, string field, int value, bool convert = false)
         {
             // Guard parameters
             if (GuardNullParam(inputObject, nameof(inputObject), out AssertResult result) ||
@@ -333,7 +382,7 @@ namespace PSRule.Runtime
                 GuardField(inputObject, field, false, out object fieldValue, out result))
                 return result;
 
-            if (CompareNumeric(fieldValue, value, out int compare, out object actual))
+            if (CompareNumeric(fieldValue, value, convert, out int compare, out object actual))
                 return compare <= 0 ? Pass() : Fail(ReasonStrings.LessOrEqual, actual, value);
 
             return Fail(ReasonStrings.Compare, fieldValue, value);
@@ -547,9 +596,9 @@ namespace PSRule.Runtime
             return false;
         }
 
-        private static bool TryInt(object obj, out int value)
+        private static bool TryInt(object obj, bool convert, out int value)
         {
-            if (obj is int ivalue)
+            if (obj is int ivalue || (convert && obj is string s && int.TryParse(s, out ivalue)))
             {
                 value = ivalue;
                 return true;
@@ -558,9 +607,9 @@ namespace PSRule.Runtime
             return false;
         }
 
-        private static bool TryLong(object obj, out long value)
+        private static bool TryLong(object obj, bool convert, out long value)
         {
-            if (obj is long lvalue)
+            if (obj is long lvalue || (convert && obj is string s && long.TryParse(s, out lvalue)))
             {
                 value = lvalue;
                 return true;
@@ -569,9 +618,9 @@ namespace PSRule.Runtime
             return false;
         }
 
-        private static bool TryFloat(object obj, out float value)
+        private static bool TryFloat(object obj, bool convert, out float value)
         {
-            if (obj is float fvalue)
+            if (obj is float fvalue || (convert && obj is string s && float.TryParse(s, out fvalue)))
             {
                 value = fvalue;
                 return true;
@@ -602,24 +651,30 @@ namespace PSRule.Runtime
             return false;
         }
 
-        private static bool CompareNumeric(object obj, int value, out int compare, out object actual)
+        private static bool CompareNumeric(object obj, int value, bool convert, out int compare, out object actual)
         {
-            if (TryInt(obj, out int iactual) || TryStringLength(obj, out iactual) || TryArrayLength(obj, out iactual))
+            if (TryInt(obj, convert, out int iactual))
             {
                 compare = iactual.CompareTo(value);
                 actual = iactual;
                 return true;
             }
-            if (TryLong(obj, out long lactual))
+            if (TryLong(obj, convert, out long lactual))
             {
                 compare = lactual.CompareTo(value);
                 actual = lactual;
                 return true;
             }
-            if (TryFloat(obj, out float factual))
+            if (TryFloat(obj, convert, out float factual))
             {
                 compare = factual.CompareTo(value);
                 actual = factual;
+                return true;
+            }
+            if (TryStringLength(obj, out iactual) || TryArrayLength(obj, out iactual))
+            {
+                compare = iactual.CompareTo(value);
+                actual = iactual;
                 return true;
             }
             compare = 0;
@@ -642,13 +697,26 @@ namespace PSRule.Runtime
         /// <summary>
         /// Fails of the value is null or empty.
         /// </summary>
-        /// <returns>Returns true if the field does not exist.</returns>
+        /// <returns>Returns true if the value is null or an empty string.</returns>
         /// <remarks>
         /// Reason: The parameter '{0}' is null or empty.
         /// </remarks>
         private bool GuardNullOrEmptyParam(string value, string parameterName, out AssertResult result)
         {
             result = string.IsNullOrEmpty(value) ? Fail(ReasonStrings.NullOrEmptyParameter, parameterName) : null;
+            return result != null;
+        }
+
+        /// <summary>
+        /// Fails of the value is null or empty.
+        /// </summary>
+        /// <returns>Returns true if the value is null or is empty.</returns>
+        /// <remarks>
+        /// Reason: The parameter '{0}' is null or empty.
+        /// </remarks>
+        private bool GuardNullOrEmptyParam(Array value, string parameterName, out AssertResult result)
+        {
+            result = value == null || value.Length == 0 ? Fail(ReasonStrings.NullOrEmptyParameter, parameterName) : null;
             return result != null;
         }